Lets give the conversations in the sidebar ellipses truncation. Any thing after 25 characters we show the ellippses.
Also I want to show last active date for each conversation in the sidebar. Show as small subtext also next to it if there are attached documents within the chat we need to show an icon that shows this. So the user can easily identify if there is a document.
So we need to make sure we are saving this and tracking this in our backend database.

Also not sure what the "Search everything" search box does but it doesnt work