When we upload the file I need an activity indicator. I want to show the upload as a thumnail just lke the image attached.
I told you the upload MUST work. There is no exception to this rule. We have a backend database. The upload must be uploaded to that database. This is critical.

I expect to be able to see all documents uploaded in the Documents tools tab. So you must design this page so that we can see the documents loaded from in there.

When we see the documents listed we should be able to have the ability select the document and chat abot that document. So you need to device a way to do this.