Lets left align the document icon and right align the date timestamp with They need to be separated with space in between the two.
Make sure that our conversations container card and the content in it has more padding between the borders and edges of the container. give the background a slight gradient for the selected conversation. so we can clearly see separation between each conversation list item. Lets use the truncate logic and show the ellipses at 18 characters.

The user should be able to hover over the title "DESKTOP WEB ONLY" and see the full title of the conversation.

Lastly I dont see our popover icon anymore so there is no way to rename the conversation nor delete the conversation

Th